pyproject.toml
[build-system]
    requires = ["setuptools>=59.7", "wheel"]
    build-backend = "setuptools.build_meta"

[tool.setuptools]
    packages = ["helpers", "monitoring"]

[project]
    name = "scriptengine-tasks-ecearth"
    version = "0.6.3"
    authors = [
        { name = "Valentina Schueller", email = "valentina.schueller@gmail.com" },
        { name = "Uwe Fladrich", email = "uwe.fladrich@protonmail.com" },
    ]
    description = "ScriptEngine tasks for use with the EC-Earth climate model"
    readme = "README.md"
    classifiers = [
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
        "Operating System :: OS Independent",
    ]
    requires-python = ">=3.8"
    dependencies = [
        "scriptengine>=0.8.1",
        "pyYAML>=5.1",
        "matplotlib>=3.1",
        "numpy>=1.18",
        "imageio>=2.18",
        "scitools-iris>=3.5",
        "cartopy>=0.20",
        "python-redmine",
    ]

    [project.urls]
        "Homepage" = "https://github.com/uwefladrich/scriptengine-tasks-ecearth"
        "Bug Tracker" = "https://github.com/uwefladrich/scriptengine-tasks-ecearth/issues"

    [project.entry-points."scriptengine.tasks"]
        "ece.mon.scalar" = "monitoring.scalar:Scalar"
        "ece.mon.timeseries" = "monitoring.timeseries:Timeseries"
        "ece.mon.diskusage_rte_scalar" = "monitoring.diskusage_rte_scalar:DiskusageRteScalar"
        "ece.mon.simulatedyears_rte_scalar" = "monitoring.simulatedyears_rte_scalar:SimulatedyearsRteScalar"
        "ece.mon.nemo_global_mean_year_mean_timeseries" = "monitoring.nemo_timeseries:NemoGlobalMeanYearMeanTimeseries"
        "ece.mon.nemo_global_sum_year_mean_timeseries" = "monitoring.nemo_timeseries:NemoGlobalSumYearMeanTimeseries"
        "ece.mon.nemo_year_mean_timeseries" = "monitoring.nemo_timeseries:NemoYearMeanTimeseries"
        "ece.mon.nemo_all_mean_map" = "monitoring.nemo_all_mean_map:NemoAllMeanMap"
        "ece.mon.nemo_month_mean_temporalmap" = "monitoring.nemo_time_mean_temporalmap:NemoMonthMeanTemporalmap"
        "ece.mon.nemo_year_mean_temporalmap" = "monitoring.nemo_time_mean_temporalmap:NemoYearMeanTemporalmap"
        "ece.mon.si3_hemis_sum_month_mean_timeseries" = "monitoring.si3_hemis_sum_month_mean_timeseries:Si3HemisSumMonthMeanTimeseries"
        "ece.mon.si3_hemis_point_month_mean_all_mean_map" = "monitoring.si3_hemis_point_month_mean_all_mean_map:Si3HemisPointMonthMeanAllMeanMap"
        "ece.mon.si3_hemis_point_month_mean_temporalmap" = "monitoring.si3_hemis_point_month_mean_temporalmap:Si3HemisPointMonthMeanTemporalmap"
        "ece.mon.oifs_all_mean_map" = "monitoring.oifs_all_mean_map:OifsAllMeanMap"
        "ece.mon.oifs_year_mean_temporalmap" = "monitoring.oifs_year_mean_temporalmap:OifsYearMeanTemporalmap"
        "ece.mon.oifs_global_mean_year_mean_timeseries" = "monitoring.oifs_global_mean_year_mean_timeseries:OifsGlobalMeanYearMeanTimeseries"
        "ece.mon.presentation.markdown" = "monitoring.markdown:Markdown"
        "ece.mon.presentation.redmine" = "monitoring.redmine:Redmine"
        "ece.mon.linear_combination" = "monitoring.linear_combination:LinearCombination"
